add increase position

+++ src/core/interface/IPriceFeed.sol
@@ -4,6 +4,6 @@ pragma solidity ^0.8.17;
 interface IPriceFeed {
     function shouldUpdate() external view returns (bool);
     function updatePrice() external;
-    function getRoundId() external view returns (uint80);
-    function getLatestAnswer() external view returns (int256);
+    function getRoundId() external view returns (uint256);
+    function getLatestAnswer() external view returns (uint256);
 }
